feat: divide util.js
