+++ src/data/film.js
@@ -1,4 +1,5 @@
-import {getRandomInteger, getRandomDecimalNumber, generateRandomNameFromArray, generateRandomDayFrom, generateRandomArrayFromArray} from '../util.js';
+import {getRandomInteger, getRandomDecimalNumber, generateRandomNameFromArray, generateRandomArrayFromArray} from '../utils/common.js';
+import {generateRandomDayFrom} from '../utils/task.js';
 import {TITLES, DIRECTORS, WRITERS, ACTORS, GENRES, COUNTRIES, POSTERS, DESCRIPTIONS, MAX_RATING, MAX_AGE_RATING, MAX_ARRAY_LENGTH, MIN, MAX_YEAR_GAP, MIN_RUNTIME, MAX_RUNTIME} from '../const.js';
 
 let counter = 0;
